requirements.txt
# LureNet Enterprise v2.0 - Professional Honeypot Platform
# Python 3.8-3.13 Required

# ============================================================================
# Web Framework & Server
# ============================================================================
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
gevent==24.2.1
simple-websocket==1.1.0
python-socketio==5.11.2

# ============================================================================
# Database & ORM
# ============================================================================
SQLAlchemy==2.0.36

# ============================================================================
# Security & Authentication
# ============================================================================
cryptography==43.0.3
bcrypt==4.2.1
passlib==1.7.4
PyJWT==2.10.1

# ============================================================================
# Configuration & CLI
# ============================================================================
PyYAML==6.0.2
click==8.1.8
python-dotenv==1.0.1

# ============================================================================
# Protocol Implementations
# ============================================================================
# SSH
paramiko==3.5.0
# FTP
pyftpdlib==2.0.1
# SMTP
aiosmtpd==1.4.6
# DNS
dnspython==2.7.0
dnslib==0.9.25
# SMB (optional)
pysmb==1.2.9.1

# ============================================================================
# Threat Intelligence & Analysis
# ============================================================================
# GeoIP
geoip2==4.8.1
maxminddb==2.6.2
# Malware Analysis
yara-python==4.5.1
pefile==2024.8.26
# IP Reputation
requests==2.32.3

# ============================================================================
# Data Processing & Utilities
# ============================================================================
python-dateutil==2.9.0
pytz==2024.2
tabulate==0.9.0
colorama==0.4.6

# ============================================================================
# System Monitoring
# ============================================================================
psutil==6.1.0

# ============================================================================
# Testing & Quality
# ============================================================================
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0
